refactor: Refactor api and output components
