+++ src/api/game-api.ts
@@ -12,33 +12,13 @@ import {
     buildRevertFunction,
     scrubAgents,
     StaticAgentRegistry
-} from "./agents";
-import { HookManager } from "./api-hooks";
-import { GameMetadata, GameOptions, MetadataManager } from "./config";
-import { ContextManager } from "./context-manager";
-import { RegalError } from "./error";
-import GameInstance from "./game-instance";
-import { GameOutput } from "./output";
-
-/**
- * Interface for the response of every `Game` API method,
- * containing some output and a `GameInstance`, if applicable.
- */
-export interface GameResponse {
-    /**
-     * The new instance state of the game.
-     *
-     * Will not be defined if an error occurred during the request,
-     * or if `getMetdataCommand` was called.
-     */
-    instance?: GameInstance;
-
-    /**
-     * The output generated by the request, which will vary in
-     * structure depending on the request and if it was successful.
-     */
-    output: GameOutput;
-}
+} from "../agents";
+import { GameMetadata, GameOptions, MetadataManager } from "../config";
+import { ContextManager } from "../context-manager";
+import { RegalError } from "../error";
+import GameInstance from "../game-instance";
+import { HookManager } from "./api-hook-manager";
+import { GameResponse, GameResponseOutput } from "./game-response";
 
 /**
  * Throws an error if `instance` or any of its properties are undefined.
@@ -96,7 +76,7 @@ const buildLogResponse = (
     let response: GameResponse;
 
     if (err !== undefined) {
-        const output: GameOutput = {
+        const output: GameResponseOutput = {
             error: err,
             wasSuccessful: false
         };
@@ -104,7 +84,7 @@ const buildLogResponse = (
             output
         };
     } else {
-        const output: GameOutput = {
+        const output: GameResponseOutput = {
             log: newInstance.output.lines,
             wasSuccessful: true
         };
