+++ src/api/index.ts
@@ -1,0 +1,8 @@
+export { GameResponse, GameResponseOutput } from "./game-response";
+export { Game } from "./game-api";
+export {
+    onPlayerCommand,
+    onStartCommand,
+    onBeforeUndoCommand
+} from "./func/api-hooks";
+export { HookManager } from "./api-hook-manager";
