+++ src/output/impl/index.ts
@@ -1,0 +1,1 @@
+export { buildInstanceOutput } from "./instance-output-impl";
