+++ src/output/index.ts
@@ -1,0 +1,4 @@
+export { recycleInstanceOutput } from "./func/recycle-instance-output";
+export { buildInstanceOutput } from "./impl";
+export { InstanceOutput } from "./instance-output";
+export { OutputLine, OutputLineType } from "./output-line";
