+++ src/output/instance-output.ts
@@ -1,0 +1,62 @@
+import GameInstance from "../game-instance";
+import { OutputLine, OutputLineType } from "./output-line";
+
+/**
+ * Manager for output in a `GameInstance`.
+ *
+ * Contains API for game developers to emit output.
+ */
+export interface InstanceOutput {
+    /** The number of `OutputLine`s that have been generated by the `GameInstance`. */
+    readonly lineCount: number;
+
+    /** The `GameInstance` that owns this `InstanceOutput`. */
+    game: GameInstance;
+
+    /** The `OutputLine`s generated during the current game cycle. */
+    lines: OutputLine[];
+
+    /**
+     * Writes a single line of output to the client.
+     *
+     * @param line The text string to be emitted.
+     * @param lineType The line's semantic meaning. (Defaults to `OutputLineType.NORMAL`)
+     */
+    writeLine(line: string, lineType?: OutputLineType): void;
+
+    /**
+     * Writes one or more lines of type `OutputLineType.NORMAL` to the output.
+     * @param lines The text to be emitted.
+     */
+    write(...lines: string[]): void;
+
+    /**
+     * Writes one or more lines of type `OutputLineType.NORMAL` to the output.
+     * @param lines The text to be emitted.
+     */
+    writeNormal(...lines: string[]): void;
+
+    /**
+     * Writes one or more lines of type `OutputLineType.MAJOR` to the output.
+     * @param lines The text to be emitted.
+     */
+    writeMajor(...lines: string[]): void;
+
+    /**
+     * Writes one or more lines of type `OutputLineType.MINOR` to the output.
+     * @param lines The text to be emitted.
+     */
+    writeMinor(...lines: string[]): void;
+
+    /**
+     * Writes one or more lines of type `OutputLineType.DEBUG` to the output.
+     * @param lines The text to be emitted.
+     */
+    writeDebug(...lines: string[]): void;
+
+    /**
+     * Writes a line of type of `OutputLineType.SECTION_TITLE` to the output.
+     * @param line The text to be emitted.
+     */
+    writeTitle(line: string): void;
+}
